Fabric8-Stack-Analysis - NPM Module
What this commit has?

1. This commit removes all the previous files, places only those needed for the NPM.
2. Has a folder Stack that has the total application.
3. Exposes index.ts that acts as the starting point that includes and exports the Parent module.
4. Added Typings.json, Tsconfig.json

How this will affect?

Since all the previous files are removed and it retains only the NPM specific modules, this application can't be brought up on its own.
It is no more a standalone application and hence it needs some application to run on.

Package entries:
Name: fabric8-stack-analysis
Version: 1.0.0

Tracks #51.